extract and abstract
