+++ src/databases/EdgeDatabase.php
@@ -1,0 +1,48 @@
+<?php
+
+namespace DominionEnterprises\NetAcuity\Databases;
+
+use GuzzleHttp\ClientInterface;
+
+/**
+ * Formats and returns our expected data set according to the Net Acuity Edge type database.
+ */
+final class EdgeDatabase extends AbstractNetAcuityDatabase
+{
+    /**
+     * EdgeDatabase constructor.
+     *
+     * @param ClientInterface $client       The injected GuzzleHttp Client.
+     * @param string          $apiUserToken The Net Acuity API User Token.
+     */
+    public function __construct(ClientInterface $client, string $apiUserToken)
+    {
+        parent::__construct($client, $apiUserToken);
+
+        $this->_databaseIdentifier = 4;
+
+        $this->_translations = [
+            'area-code' => 'edge-area-codes',
+            'city' => 'edge-city',
+            'city-code' => 'edge-city-code',
+            'city-conf' => 'edge-city-conf',
+            'conn-speed' => 'edge-conn-speed',
+            'continent-code' => 'edge-continent-code',
+            'country' => 'edge-country',
+            'country-code' => 'edge-country-code',
+            'country-conf' => 'edge-country-conf',
+            'gmt-offset' => 'edge-gmt-offset',
+            'in-dist' => 'edge-in-dst',
+            'latitude' => 'edge-latitude',
+            'longitude' => 'edge-longitude',
+            'metro-code' => 'edge-metro-code',
+            'postal-conf' => 'edge-postal-conf',
+            'region' => 'edge-region',
+            'region-code' => 'edge-region-code',
+            'region-conf' => 'edge-region-conf',
+            'timezone-name' => 'edge-timezone-name',
+            'two-letter-country' => 'edge-two-letter-country',
+            'zip-code' => 'edge-postal-code',
+        ];
+    }
+}
